fix: font line-height scale error
the em unit is related to parent font size, but in calculating the
width, it is related to the current font size. that means
calc(var(-bp-xlarge) * 1em) are multiple current element font size.
when font size is large, like 100px, it gove you a huge line
height. I change 1 em back to 16px.